// src/models/IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionStatus/types.ts
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionConditionModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionCondition } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionCondition';
import { IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNamesModelOptions as IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames } from '../IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames';

export interface IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionStatusModelOptions {
  acceptedNames?: IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionNames;
  conditions?: Array<IoK8sApiextensionsApiserverPkgApisApiextensionsV1CustomResourceDefinitionCondition>;
  storedVersions?: Array<string>;
}